# Make AI faster by speeding up some combat code.