--- a/game-app/game-core/src/main/java/games/strategy/triplea/delegate/power/calculator/BombardmentCombatValue.java
+++ b/game-app/game-core/src/main/java/games/strategy/triplea/delegate/power/calculator/BombardmentCombatValue.java
@@ -111,7 +111,7 @@ public class BombardmentCombatValue implements CombatValue {
   static class BombardmentStrength implements StrengthCalculator {
 
     int gameDiceSides;
-    Collection<TerritoryEffect> territoryEffects;
+    @Nonnull Collection<TerritoryEffect> territoryEffects;
     AvailableSupports supportFromFriends;
     AvailableSupports supportFromEnemies;