--- a/game-app/game-core/src/main/java/games/strategy/triplea/delegate/power/calculator/SupportCalculator.java
+++ b/game-app/game-core/src/main/java/games/strategy/triplea/delegate/power/calculator/SupportCalculator.java
@@ -63,21 +63,23 @@ public class SupportCalculator {
       final Predicate<Unit> canSupport =
           Matches.unitIsOfType((UnitType) rule.getAttachedTo())
               .and(Matches.unitIsOwnedByAnyOf(rule.getPlayers()));
-      final List<Unit> supporters = CollectionUtils.getMatches(unitsGivingTheSupport, canSupport);
-      if (supporters.isEmpty()) {
-        continue;
-      }
-      final List<Unit> impArtTechUnits =
-          rule.getImpArtTech()
-              ? CollectionUtils.getMatches(
-                  supporters, Matches.unitOwnerHasImprovedArtillerySupportTech())
-              : List.of();
-
+      final Predicate<Unit> impArtTech =
+          rule.getImpArtTech() ? Matches.unitOwnerHasImprovedArtillerySupportTech() : u -> false;
       final IntegerMap<Unit> unitsForRule = new IntegerMap<>();
-      supporters.forEach(unit -> unitsForRule.put(unit, rule.getNumber()));
-      impArtTechUnits.forEach(unit -> unitsForRule.add(unit, rule.getNumber()));
-      supportUnits.put(rule, unitsForRule);
-      supportRules.computeIfAbsent(rule.getBonusType(), (bonusType) -> new ArrayList<>()).add(rule);
+      for (Unit unit : unitsGivingTheSupport) {
+        if (!canSupport.test(unit)) {
+          continue;
+        }
+
+        unitsForRule.put(unit, rule.getNumber());
+        if (impArtTech.test(unit)) {
+          unitsForRule.add(unit, rule.getNumber());
+        }
+      }
+      if (!unitsForRule.isEmpty()) {
+        supportUnits.put(rule, unitsForRule);
+        supportRules.computeIfAbsent(rule.getBonusType(), (bt) -> new ArrayList<>()).add(rule);
+      }
     }
   }