## Change Summary & Additional Notes
This PR speeds up several areas in the code that were showing in performance profiles.

With this change, we achieve the following speed ups on a given AI's turn on "Another World" map:

1. Main game thread CPU time goes from 90s to 70s
2. Each of 15 (on my machine) AI combat simulation threads go from 73s CPU time to 52s CPU time

This is with this save game: [speed_test.zip](https://github.com/triplea-game/triplea/files/9137146/speed_test.zip)
Before starting it, set player "Malliton" to Hard AI.

The above profiling results were measured with VisualVM, attaching to the game for the entire duration of the AI turn.
<!--
- If multiple commits, summarize what has changed
- Mention any manual testing done.
- If there are UI updates, please include before & after screenshots
-->

## Release Note
<!--
Include a release note if there is a bug fix or a visible change for players.
For format & syntax help, see:
https://github.com/triplea-game/triplea/blob/master/docs/development/reference/pr-release-notes.md
-->

<!--RELEASE_NOTE--><!--END_RELEASE_NOTE-->